Add typescript declarations for core-js-compat
Reuse it for the core-js-builder declarations to prevent duplicated code
