+++ packages/core-js-compat/package.json
@@ -19,6 +19,7 @@
   },
   "sideEffects": false,
   "main": "index.js",
+  "types": "index.d.ts",
   "dependencies": {
     "browserslist": "^4.21.5"
   }
